refactor(progression) : move evidence factory to ProgressionEvidence and fix ratio computation

+++ src/main/java/eu/nicosworld/rithmo/engine/capture/justification/GeometricJustification.java
@@ -1,3 +1,3 @@
 package eu.nicosworld.rithmo.engine.capture.justification;
 
-public record GeometricJustification(int ratio) implements ProgressionEvidence {}
+public record GeometricJustification(double ratio) implements ProgressionEvidence {}
